python-ai/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
Pillow
groq
pdf2image